Ditch `mktmpdir` for `mktemp -d`
As of b515fe10e17beeffe25eaf76752bb541a6a30f78 (merged in
351bbf3895e9ca0ef9f3126fc1f3b0425f76ca20, 2019-03-26)), the generated
install script now remembers to clean up its temporary subdirectory
when it exits.

However, the `tmpdir` in `rm -rf "${tmpdir}` is set by a function from
shlib called `mktmpdir`.  `mktmpdir` has wildly different behavior
when `TMPDIR` is set compared to when it's not set:

    mktmpdir() {
      test -z "$TMPDIR" && TMPDIR="$(mktemp -d)"
      mkdir -p "${TMPDIR}"
      echo "${TMPDIR}"
    }

* If `TMPDIR` is not already set: it is *necessary* to clean up the
  returned directory when the program is done; it has created a new
  temporary directory that would otherwise be leaked.
* If `TMPDIR` is already set: it is *impermissible* to clean up the
  returned directory when the program is done; it returns the shared
  temporary directory, other programs may be using it, the user may
  not have permission to remove it.

This means that if the environment variable `TMPDIR` is already set
when the install script is run, then the install script will remove
(or attempt to remove) the entire *shared* temporary directory.

The most obvious solution is to just use `mktemp -d` directly.  So do
that.