ci: [Proposal] Introduce reviewdog/action-golangci-lint
This PR can be follow-up of #167.
PR #167 will stop using all linters, and I propose to introduce
[reviewdog/action-golangci-lint](https://github.com/reviewdog/action-golangci-lint)
as one of alternative options to run all linters.

It runs golangci-lint on Pull Requests and report results in changed
lines. (Note that for PR from forked repository, it cannot post comments
and it will use log console to report result though because of
permission issue [1]).

It's just an proposal, so please feel to close it if you don't like it
or ask any questions.

[1]: https://github.com/reviewdog/reviewdog#option-1-run-reviewdog-from-github-actions-w-secretsgithub_token-experimental